Cambios generales de ayuda y metadata. Opciones de p!decir rehechas

+++ commands/Pure/decir.js
@@ -12,31 +12,30 @@ module.exports = {
         'common'
     ],
     options: [
-
+        '`-d` o `--delete` para borrar tu mensaje'
     ],
 	
 	execute(message, args) {
         if(args.length > 0) {
-            let sentence = '';
-            for(var i = 0; i < args.length; i++) {
-                switch(args[i]) {
-                    case '-d':
-                        message.delete();
-                        break;
-
-                    case '--del':
-                        message.delete();
-                        break;
-                    
-                    default:
-                        sentence += ' ' + args[i];
-                        break;
-                }
-            }
+            const sentence = args.join(' ');
+            let dflag = false;
+            
+            args.some((arg, i) => {
+				if(arg.startsWith('--'))
+					switch(arg.slice(2)) {
+					case 'del': dflag = true; break;
+					case 'delete': dflag = true; break;
+					}
+				else if(arg.startsWith('-'))
+					for(c of arg.slice(1))
+						switch(c) {
+						case 'd': dflag = true; break;
+						}
+			});
+            if(dflag) message.delete();
 
             const minus = sentence.toLowerCase();
-
-            if(message.channel.guild.id === '654471968200065034' && minus.indexOf('hourai') !== -1 && minus.indexOf('hourai doll') !== minus.indexOf('hourai') && minus.indexOf('houraidoll') === -1)
+            if(message.channel.guild.id === global.serverid.hourai && minus.indexOf('hourai') !== -1 && minus.indexOf('hourai doll') !== minus.indexOf('hourai') && minus.indexOf('houraidoll') === -1)
                 message.channel.send('No me hagai decir weas de hourai, ¿yapo? Gracias <:haniwaSmile:659872119995498507>');
             else message.channel.send(sentence);
         } else message.channel.send(':warning: tienes que especificar lo que quieres que diga.');
